{"schema":"decodex/commit/1","summary":"Split MCP test support modules","authority":"manual"}
